# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HEXA Purchase Assistant API
  version: 0.1.0
  description: |
    FastAPI backend for Flutter app, WhatsApp webhook, and Super Admin.
    AI is used for parsing only; business logic stays server-side.
servers:
  - url: https://api.example.com
    description: Production
  - url: http://localhost:8000
    description: Local
...
paths:
...
  /v1/businesses/{businessId}/suppliers:
    get:
      tags: [contacts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List suppliers
    post:
      tags: [contacts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreate'
      responses:
        '201':
          description: Created
  /v1/businesses/{businessId}/brokers:
    get:
      tags: [contacts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List brokers
    post:
      tags: [contacts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '201':
          description: Created broker
  /v1/businesses/{businessId}/brokers/{brokerId}:
    get:
      tags: [contacts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: brokerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Broker
  /v1/businesses/{businessId}/suppliers/{supplierId}:
    get:
      tags: [contacts]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: businessId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: supplierId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Supplier
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SupplierCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        location:
          type: string
        broker_id:
          type: string
          format: uuid